SignIn Screen ready with cleaner code & stylesheet
