+++ src/router.js
@@ -2,10 +2,7 @@ import React from 'react';
 import firebase from 'firebase/app';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import RankingScreen from './screens/RankingScreen';
-import AddInfoScreen from './screens/AddInfoScreen';
 import SignInScreen from './screens/SignInScreen';
-import HomeScreen from './screens/HomeScreen';
 import firebaseConfig from './firebase';
 
 if (!firebase.apps.length) {
@@ -20,27 +17,6 @@ const Routes = () => (
       <Stack.Screen
         name="SignInScreen"
         component={SignInScreen}
-        options={{
-          headerShown: false,
-        }}
-      />
-      <Stack.Screen
-        name="HomeScreen"
-        component={HomeScreen}
-        options={{
-          headerShown: false,
-        }}
-      />
-      <Stack.Screen
-        name="AddInfoScreen"
-        component={AddInfoScreen}
-        options={{
-          headerShown: false,
-        }}
-      />
-      <Stack.Screen
-        name="RankingScreen"
-        component={RankingScreen}
         options={{
           headerShown: false,
         }}
